Develop WFViewer module to show board-by-board incoming data
A few major changes to artdaq-demo in this commit:

WFViewer has been modified s.t. when the "-m on" option is chosen
during the "init" transition, once "start" is called, two windows pop
up. The first window shows the distribution of ADC counts for each
board. The second window shows, for a single event, a snapshot of all
the ADC values for that event delivered by the board.

Additionally, each boardreader's FHiCL gets a different random number
generator seed -- this had not previously been the case, so initially
I was seeing identical data from each boardreader.

Also, I've added logic so that now the DemoControl.rb script is able
to "think" in terms of boards containing multiple fragments; therefore
I've added a "fragments_per_board" variable to the FHiCL string used
to initialize the boardreaders. Right now, rather than being a
settable command line value as I plan to make it, it's simply a
variable declared and assigned on line 852. ALSO, in this commit,
please leave it set to 1 -- work on this feature is ongoing.