Add # of fragments per board per event as variable; improve on the GUI
Two major changes. One is that I completed the implementation of
artdaq-demo's ability to allow you spread an event across multiple
boards and fragments. DemoControl.rb now takes a "fragments-per-board"
argument which sets this; in the manage2x2x2System.sh script this is
set to 2, but can easily be changed. If the argument isn't supplied,
DemoControl.rb defaults to 1 fragment per board.

Also, the GUI has been improved; the primary changes are improved
labeling of the plots, and that the histograms now contain data for
all events, not just those which pass the prescale. Also, if
"digital_sum_only" is set to "true", then only the histograms, and not
the single-event graphs, will be displayed.