Added the "CreateNewFragmentGenerator.rb" script
This script, added to the tools/ directory, is designed to streamline
the process of creating new fragment overlays / fragment generators by
taking the desired name for the overlay and generator from the user
and then copying the Toy* source files, renaming all tokens in the
files accordingly, and adding the new classes to the build
scripts. It's left to the user to actually change the implementation
details of the classes so they differ from the Toy* implementation.