Improved modularization of the Ruby FHiCL-configuration code
Basically, generateToy.rb and generateV1720.rb generated code not just
for the ToySimulator and V172xSimulator fragment generators, but also
for the BoardReaderMain app. Therefore (A) they both contained
identical code, and (B) there wasn't a FHiCL-code generating function
available if you wanted to work with ToySimulator or V172xSimulator
standalone. That's been fixed; the prior roles of generateToy and
generateV172x have been taken over by generateBoardReaderMain, which
takes as one of its arguments the generator code, about which it knows
nothing.