Minor interface and logic changes to some generate* functions
Decision on whether to read a parameter in from a *.fcl file in the
generate* functions now depends explicitlywhether that parameter is
"nil" or not (previously a simple boolean test, so if a user gave a
parameter a value of "false" this would be ignored and its value read
in from a file!). Also, this feature, previously not existent in
generateWFViewer.rb, has been added.