Add "throttle_usecs" parameter to ToySimulator
Some investigation revealed that the MPI_Send error which sometimes
appeared when running the DAQ with toy fragments went away when an
adjustable pause was added at the beginning of the
ToySimulator::getNext_() function; motivated by Kurt's added pause to
V172xSimulator, this adjustment is controlled via the new
throttle_usecs parameter, which can be set either in the generateToy()
argument list, or, failing that, in the tools/fcl/ToySimulator.fcl
file, where it defaults to 100000 (i.e., 0.1 s). Correspondingly, the
prescale default in WFViewer.fcl has been changed from 1000 to 200,
since the pause gives Root more time to paint its canvases.