Various improvements to the WFViewer module; simplification of control scripts
One improvement to the WFViewer module is that it now no longer takes
the "fragments_per_board" and "fragment_receiver_count" variables used
to arrange the plots within a canvas (fragments_per_board was # of
plots on the x-axis, fragment_receiver_count, y-axis). This is because
"fragments_per_board", put in just after New Year's, has been
eliminated from the Ruby control scripts, as it turned out to be
unnecessary and redundant for associating more than one fragment
generator with a process (the composite driver, available in artdaq
v1_05_08, does this). Instead, it automatically calculates the layout
for you, although you can override this by setting the new
"num_x_plots" and "num_y_plots" variables.